+++ ribs/emitters/opt/__init__.py
@@ -1,6 +1,9 @@
 """Internal subpackage with optimizers for use across emitters."""
 from ribs.emitters.opt._cma_es import CMAEvolutionStrategy
+from ribs.emitters.opt._gradients import AdamOpt, GradientAscentOpt
 
 __all__ = [
     "CMAEvolutionStrategy",
+    "AdamOpt",
+    "GradientAscentOpt",
 ]