+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 #### API
 
+- Add `GradientAborescenceEmitter`, which is used to implement CMA-MEGA (#240)
 - Update emitter `tell()` docstrings to no longer say "Inserts entries into archive" (#247)
 - Expose `emitter.restarts` as a property (#248)
 - Specify that `x0` is 1D for all emitters (#244)